UI changes to show decoding data for permits
