+++ ui/pages/confirmations/components/confirm/info/typed-sign/permit-simulation/decoded-simulation/decoded-simulation.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const DecodedSimulation: React.FC<object> = () => (
+  <div>"DECODED SIMULATION IMPLEMENTATION"</div>
+);
+
+export default DecodedSimulation;
