+++ ui/pages/confirmations/components/confirm/info/typed-sign/permit-simulation/decoded-simulation/index.ts
@@ -1,0 +1,1 @@
+export { default as DecodedSimulation } from './decoded-simulation';
