+++ ui/pages/confirmations/components/confirm/info/typed-sign/permit-simulation/default-simulation/index.ts
@@ -1,0 +1,1 @@
+export { default as DefaultSimulation } from './default-simulation';
