+++ ui/pages/confirmations/components/confirm/info/typed-sign/permit-simulation/permit-simulation.test.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import { act } from 'react-dom/test-utils';
+import { waitFor } from '@testing-library/dom';
 
 import { getMockTypedSignConfirmStateForRequest } from '../../../../../../../../test/data/confirmations/helper';
 import { renderWithConfirmContextProvider } from '../../../../../../../../test/lib/confirmations/render-helpers';
-import {
-  permitNFTSignatureMsg,
-  permitSignatureMsg,
-} from '../../../../../../../../test/data/confirmations/typed_sign';
+import { permitSignatureMsg } from '../../../../../../../../test/data/confirmations/typed_sign';
 import { memoizedGetTokenStandardAndDetails } from '../../../../../utils/token';
 import PermitSimulation from './permit-simulation';
 
@@ -27,8 +25,12 @@ describe('PermitSimulation', () => {
     memoizedGetTokenStandardAndDetails?.cache?.clear?.();
   });
 
-  it('renders component correctly', async () => {
-    const state = getMockTypedSignConfirmStateForRequest(permitSignatureMsg);
+  it('should render default simulation if decoding api does not return result', async () => {
+    const state = getMockTypedSignConfirmStateForRequest({
+      ...permitSignatureMsg,
+      decodingLoading: false,
+      decodingData: undefined,
+    });
     const mockStore = configureMockStore([])(state);
 
     await act(async () => {
@@ -42,8 +44,18 @@ describe('PermitSimulation', () => {
     });
   });
 
-  it('renders correctly for NFT permit', async () => {
-    const state = getMockTypedSignConfirmStateForRequest(permitNFTSignatureMsg);
+  it('should render default simulation if decoding api returns error', async () => {
+    const state = getMockTypedSignConfirmStateForRequest({
+      ...permitSignatureMsg,
+      decodingLoading: false,
+      decodingData: {
+        stateChanges: null,
+        error: {
+          message: 'some error',
+          type: 'SOME_ERROR',
+        },
+      },
+    });
     const mockStore = configureMockStore([])(state);
 
     await act(async () => {
@@ -52,9 +64,28 @@ describe('PermitSimulation', () => {
         mockStore,
       );
 
-      expect(await findByText('Withdraw')).toBeInTheDocument();
-      expect(await findByText('#3606393')).toBeInTheDocument();
+      expect(await findByText('30')).toBeInTheDocument();
       expect(container).toMatchSnapshot();
+    });
+  });
+
+  it('should not render default simulation if decodingLoading is true', async () => {
+    const state = getMockTypedSignConfirmStateForRequest({
+      ...permitSignatureMsg,
+      decodingLoading: true,
+    });
+    const mockStore = configureMockStore([])(state);
+
+    await act(async () => {
+      const { container, queryByTestId } = renderWithConfirmContextProvider(
+        <PermitSimulation />,
+        mockStore,
+      );
+
+      await waitFor(() => {
+        expect(queryByTestId('30')).not.toBeInTheDocument();
+        expect(container).toMatchSnapshot();
+      });
     });
   });
 });
